Move headers into inclusion scope instead of source scope.
