+++ src/Princept.cpp
@@ -2,7 +2,7 @@
 // Created by Felix Klauke on 29.01.18.
 //
 
-#include "Princept.h"
+#include "../include/Princept.h"
 
 Princept::Princept(char **cliArguments) {
